requirements.txt
streamlit>=1.28.0
crewai>=0.11.0
crewai-tools>=0.0.6
selenium>=4.15.0
pyyaml>=6.0.1
gspread>=5.12.0
google-auth>=2.23.0